Add support for checking the AuthorityKeyIdentifier extension
This is used for working with certificates that have duplicate subject
names, where the "issuer" is not sufficiently unique to identify the
correct parent certificate.

It's hidden behind an ifdef because it breaks binary compatibility, but
I'd anticipate that the ifdef could be removed in a future release, when
this becomes the default.